Renamed and refactored seeds
