+++ src/models/group-user.model.ts
@@ -4,13 +4,9 @@ import { Application } from '../declarations'
 export default (app: Application): any => {
   const sequelizeClient: Sequelize = app.get('sequelizeClient')
   const groupUser = sequelizeClient.define('group_user', {
-    isOwner: {
-      type: DataTypes.BOOLEAN,
-      defaultValue: false
-    },
-    isInviteAccepted: {
-      type: DataTypes.BOOLEAN,
-      defaultValue: false
+    text: {
+      type: DataTypes.STRING,
+      allowNull: false
     }
   }, {
     hooks: {
@@ -20,9 +16,10 @@ export default (app: Application): any => {
     }
   });
 
-  (groupUser as any).associate = (models: any) => {
+  (groupUser as any).associate = function (models: any) {
+    (groupUser as any).belongsTo(models.user, { foreignKey: 'userId' });
     (groupUser as any).belongsTo(models.group, { foreignKey: 'groupId' });
-    (groupUser as any).belongsTo(models.user, { foreignKey: 'userId' })
+    (groupUser as any).hasOne(models.group_user_rank, { foreignKey: 'groupUserRank' })
   }
 
   return groupUser
