+++ src/models/group.model.ts
@@ -3,12 +3,16 @@ import { Application } from '../declarations'
 
 export default (app: Application): any => {
   const sequelizeClient: Sequelize = app.get('sequelizeClient')
-  const Group = sequelizeClient.define('group', {
+  const group = sequelizeClient.define('group', {
     id: {
       type: DataTypes.UUID,
       defaultValue: DataTypes.UUIDV1,
       allowNull: false,
       primaryKey: true
+    },
+    name: {
+      type: DataTypes.STRING,
+      allowNull: false
     }
   }, {
     hooks: {
@@ -18,8 +22,9 @@ export default (app: Application): any => {
     }
   });
 
-  (Group as any).associate = (models: any) => {
-    (Group as any).belongsToMany(models.user, { foreignKey: 'groupId', through: models.group_user })
+  (group as any).associate = (models: any) => {
+    (group as any).belongsToMany(models.user, { through: 'group_user' }) // user can join multiple orgs
   }
-  return Group
+
+  return group
 }
