+++ src/models/party.model.ts
@@ -3,16 +3,12 @@ import { Application } from '../declarations'
 
 export default (app: Application): any => {
   const sequelizeClient: Sequelize = app.get('sequelizeClient')
-  const organization = sequelizeClient.define('organization', {
+  const Party = sequelizeClient.define('party', {
     id: {
       type: DataTypes.UUID,
       defaultValue: DataTypes.UUIDV1,
       allowNull: false,
       primaryKey: true
-    },
-    name: {
-      type: DataTypes.STRING,
-      allowNull: false
     }
   }, {
     hooks: {
@@ -22,9 +18,8 @@ export default (app: Application): any => {
     }
   });
 
-  (organization as any).associate = (models: any) => {
-    (organization as any).belongsToMany(models.user, { through: 'organization_user' }) // user can join multiple orgs
+  (Party as any).associate = (models: any) => {
+    (Party as any).belongsToMany(models.user, { foreignKey: 'partyId', through: models.party_user })
   }
-
-  return organization
+  return Party
 }
