+++ src/models/scene.model.ts
@@ -1,4 +1,3 @@
-// TODO:
 import { Sequelize, DataTypes } from 'sequelize'
 import { Application } from '../declarations'
 
@@ -17,8 +16,6 @@ export default (app: Application): any => {
   });
 
   (scene as any).associate = (models: any) => {
-    // (scene as any).belongsTo(models.user); // or group
-    // (scene as any).belongsToMany(models.object, { through: models.scene_object })
   }
 
   return scene
