+++ src/models/user.model.ts
@@ -43,14 +43,14 @@ export default (app: Application): any => {
   (user as any).associate = (models: any) => {
     (user as any).belongsTo(models.user_role, { foreignKey: 'userRole' });
     (user as any).belongsTo(models.instance); // user can only be in one room at a time
-    (user as any).belongsTo(models.group, { through: 'group_user', foreignKey: 'userId' }); // user can only be part of one group at a time
+    (user as any).belongsTo(models.party, { through: 'party_user', foreignKey: 'userId' }); // user can only be part of one party at a time
     (user as any).hasMany(models.collection);
     (user as any).hasMany(models.entity);
     (user as any).hasMany(models.static_resource);
     (user as any).belongsToMany(models.user, { as: 'user', through: models.user_relationship });
     (user as any).belongsToMany(models.user, { as: 'relatedUser', through: models.user_relationship });
-    (user as any).belongsToMany(models.organization, { through: models.organization_user }); // user can join multiple orgs
-    (user as any).belongsToMany(models.organization_user_rank, { through: models.organization_user }); // user can join multiple orgs
+    (user as any).belongsToMany(models.group, { through: models.group_user }); // user can join multiple orgs
+    (user as any).belongsToMany(models.group_user_rank, { through: models.group_user }); // user can join multiple orgs
     (user as any).hasMany(models.identity_provider)
   }
