+++ src/services/access-control-scope/access-control-scope.seed.ts
@@ -1,0 +1,14 @@
+export const seed = {
+  disabled: false,
+  delete: true,
+  path: 'access-control-scope',
+  randomize: false,
+  templates:
+        [
+          { name: 'none' },
+          { name: 'self' },
+          { name: 'all' }
+        ]
+}
+
+export default seed
