+++ src/services/auth-management/auth-management.service.ts
@@ -1,7 +1,7 @@
 import { ServiceAddons } from '@feathersjs/feathers'
 import { Application } from '../../declarations'
 import { Authmanagement } from './auth-management.class'
-import notifier from './notifier'
+import notifier from './auth-management.notifier'
 import hooks from './auth-management.hooks'
 import authManagement from 'feathers-authentication-management'
