+++ src/services/auth-management/auth-management.utils.ts
@@ -1,4 +1,3 @@
-
 import { Application } from '../../declarations'
 
 export const getLink = (type: string, hash: string): string =>
