+++ src/services/collection-type/collection-type.seed.ts
@@ -1,0 +1,12 @@
+export const seed = {
+  disabled: false,
+  delete: true,
+  path: 'collection-type',
+  templates:
+    [
+      { name: 'scene' },
+      { name: 'inventory' }
+    ]
+}
+
+export default seed
