+++ src/services/component-type/component-type.seed.ts
@@ -1,0 +1,54 @@
+export const seed = {
+  disabled: false,
+  delete: true,
+  path: 'component-type',
+  randomize: false,
+  templates:
+      [
+        // Default Aframe components
+        { name: 'animation' },
+        { name: 'background' },
+        { name: 'camera' },
+        { name: 'cursor' },
+        { name: 'debug' },
+        { name: 'device-orientation-permission-ui' },
+        { name: 'embedded' },
+        { name: 'fog' },
+        { name: 'geometry' },
+        { name: 'gltf-model' },
+        { name: 'hand-controls' },
+        { name: 'keyboard-shortcuts' },
+        { name: 'laser-controls' },
+        { name: 'light' },
+        { name: 'line' },
+        { name: 'link' },
+        { name: 'loading-screen' },
+        { name: 'look-controls' },
+        { name: 'obj-model' },
+        { name: 'oculus-go-controls' },
+        { name: 'pool' },
+        { name: 'position' },
+        { name: 'raycaster' },
+        { name: 'renderer' },
+        { name: 'rotation' },
+        { name: 'scale' },
+        { name: 'screenshot' },
+        { name: 'shadow' },
+        { name: 'sound' },
+        { name: 'stats' },
+        { name: 'text' },
+        { name: 'tracked-controls' },
+        { name: 'visible' },
+        { name: 'vive-controls' },
+        { name: 'vive-focus-controls' },
+        { name: 'vr-mode-ui' },
+        { name: 'wasd-controls' },
+        { name: 'windows-motion-controls' },
+        // Networked Aframe
+        { name: 'networked' },
+        // Custom Components
+        { name: 'grid' }
+      ]
+}
+
+export default seed
