+++ src/services/component-type/service.ts
@@ -1,12 +1,12 @@
 import { ServiceAddons } from '@feathersjs/feathers'
 import { Application } from '../../declarations'
-import { Organization } from './organization.class'
-import createModel from '../../models/organization.model'
-import hooks from './organization.hooks'
+import { ComponentType } from './component-type.class'
+import createModel from '../../models/component-type.model'
+import hooks from './component-type.hooks'
 
 declare module '../../declarations' {
   interface ServiceTypes {
-    'organization': Organization & ServiceAddons<any>
+    'component-type': ComponentType & ServiceAddons<any>
   }
 }
 
@@ -17,9 +17,9 @@ export default (app: Application): any => {
     multi: true
   }
 
-  app.use('/organization', new Organization(options, app))
+  app.use('/component-type', new ComponentType(options, app))
 
-  const service = app.service('organization')
+  const service = app.service('component-type')
 
   service.hooks(hooks)
 }
