+++ src/services/group-user-rank/group-user-rank.class.ts
@@ -1,7 +1,7 @@
 import { Service, SequelizeServiceOptions } from 'feathers-sequelize'
 import { Application } from '../../declarations'
 
-export class Organization extends Service {
+export class GroupUserRank extends Service {
   constructor (options: Partial<SequelizeServiceOptions>, app: Application) {
     super(options)
   }
