+++ src/services/group-user-rank/group-user-rank.seed.ts
@@ -1,0 +1,13 @@
+export const seed = {
+  disabled: false,
+  delete: true,
+  path: 'group-user-rank',
+  templates:
+        [
+          { name: 'principal' },
+          { name: 'teacer' },
+          { name: 'student' }
+        ]
+}
+
+export default seed
