+++ src/services/group-user/group-user.hooks.ts
@@ -1,47 +1,13 @@
-import * as authentication from '@feathersjs/authentication'
-import { HookContext } from '@feathersjs/feathers'
-import { disallow } from 'feathers-hooks-common'
-import { BadRequest } from '@feathersjs/errors'
-
-// Don't remove this comment. It's needed to format import lines nicely.
-
-const { authenticate } = authentication.hooks
-
-// For now only Group Owner will be able to remove user
-const removeUserFromGroupQuery = () => {
-  return (context: HookContext) => {
-    Object.assign(context.params.query, {
-      isOwner: true,
-      userId: context.params.user.userId
-    })
-    return context
-  }
-}
-
-const validateGroupId = () => {
-  return async (context: HookContext) => {
-    if (!context?.params?.query?.groupId) {
-      return await Promise.reject(new BadRequest('Group Id is required!'))
-    }
-    return context
-  }
-}
 
 export default {
   before: {
-    all: [
-      authenticate('jwt'),
-      validateGroupId()
-    ],
+    all: [],
     find: [],
-    get: [disallow()],
+    get: [],
     create: [],
-    update: [disallow()],
+    update: [],
     patch: [],
-    remove: [
-      // For now only admin will be able to remove the user from group!
-      removeUserFromGroupQuery()
-    ]
+    remove: []
   },
 
   after: {
@@ -59,7 +25,7 @@ export default {
     find: [],
     get: [],
     create: [],
-    update: [disallow()],
+    update: [],
     patch: [],
     remove: []
   }
