+++ src/services/group-user/group-user.service.ts
@@ -10,7 +10,7 @@ declare module '../../declarations' {
   }
 }
 
-export default (app: Application): void => {
+export default (app: Application): any => {
   const options = {
     Model: createModel(app),
     paginate: app.get('paginate'),
