+++ src/services/group/group.hooks.ts
@@ -1,21 +1,12 @@
-import * as authentication from '@feathersjs/authentication'
-import { disallow } from 'feathers-hooks-common'
-
-// import attachOwnerIdInSavingContact from '../../hooks/set-loggedin-user-in-body'
-
-// Don't remove this comment. It's needed to format import lines nicely.
-
-const { authenticate } = authentication.hooks
 
 export default {
   before: {
-    all: [authenticate('jwt')],
+    all: [],
     find: [],
     get: [],
     create: [],
-    update: [disallow()],
+    update: [],
     patch: [],
-    // TODO: Need to ask if we allow user to remove group or not
     remove: []
   },
