+++ src/services/group/group.service.ts
@@ -10,9 +10,10 @@ declare module '../../declarations' {
   }
 }
 
-export default (app: Application): void => {
+export default (app: Application): any => {
   const options = {
     Model: createModel(app),
+    paginate: app.get('paginate'),
     multi: true
   }
