--- src/services/organization-user-rank/organization-user-rank.class.ts
@@ -1,8 +1,0 @@
-import { Service, SequelizeServiceOptions } from 'feathers-sequelize'
-import { Application } from '../../declarations'
-
-export class OrganizationUserRank extends Service {
-  constructor (options: Partial<SequelizeServiceOptions>, app: Application) {
-    super(options)
-  }
-}
