+++ src/services/resource-type/resource-type.seed.ts
@@ -1,0 +1,32 @@
+export const seed = {
+  disabled: false,
+  delete: true,
+  path: 'resource-type',
+  randomize: false,
+  templates:
+        [
+          { name: 'accessControlScope' },
+          { name: 'accessControl' },
+          { name: 'attribution' },
+          { name: 'collection' },
+          { name: 'collectionType' },
+          { name: 'component' },
+          { name: 'componentType' },
+          { name: 'entity' },
+          { name: 'entityType' },
+          { name: 'party' },
+          { name: 'party-user' },
+          { name: 'instance' },
+          { name: 'license' },
+          { name: 'location' },
+          { name: 'location-user' },
+          { name: 'group' },
+          { name: 'resourceType' },
+          { name: 'userRole' },
+          { name: 'staticResource' },
+          { name: 'staticResourceType' },
+          { name: 'user' }
+        ]
+}
+
+export default seed
