+++ src/services/static-resource-type/static-resource-type.seed.ts
@@ -1,0 +1,18 @@
+export const seed = {
+  disabled: false,
+  delete: true,
+  path: 'static-resource-type',
+  randomize: false,
+  templates: [
+    { type: 'image' },
+    { type: 'video' }, // parse metadata for video type (eg 360-eac)
+    { type: 'audio' },
+    { type: 'model3d' },
+    { type: 'script' },
+    { type: 'volumetric' }, // any volumetric file, parse metadata for type
+    { type: 'json' }, // JSON data
+    { type: 'data' } // arbitrary data of any other type
+  ]
+}
+
+export default seed
