+++ src/services/user-relationship-type/user-relationship-type.seed.ts
@@ -1,0 +1,14 @@
+export const seed = {
+  disabled: false,
+  delete: true,
+  path: 'user-relationship-type',
+  randomize: false,
+  templates:
+      [
+        { name: 'requested' }, // Default state of relatedUser
+        { name: 'friend' },
+        { name: 'blocked' }
+      ]
+}
+
+export default seed
