+++ src/services/user-role/user-role.seed.ts
@@ -1,0 +1,14 @@
+export const seed = {
+  disabled: false,
+  delete: true,
+  path: 'user-role',
+  templates:
+      [
+        { name: 'admin' },
+        { name: 'moderator' },
+        { name: 'user' },
+        { name: 'guest' }
+      ]
+}
+
+export default seed
