+++ test/services/group-user.test.ts
@@ -1,9 +1,9 @@
 import assert from 'assert'
 import app from '../../src/app'
 
-describe('\'GroupUser\' service', () => {
+describe('\'PartyUser\' service', () => {
   it('registered the service', () => {
-    const service = app.service('group-user')
+    const service = app.service('party-user')
 
     assert.ok(service, 'Registered the service')
   })
