+++ test/services/organization-user-rank.test.ts
@@ -1,9 +1,9 @@
 import assert from 'assert'
 import app from '../../src/app'
 
-describe('\'OrganizationUserRank\' service', () => {
+describe('\'GroupUserRank\' service', () => {
   it('registered the service', () => {
-    const service = app.service('organization-user-rank')
+    const service = app.service('group-user-rank')
 
     assert.ok(service, 'Registered the service')
   })
